requirements.txt
fastapi
uvicorn[standard]
torch==2.6.0
torchvision
pillow
numpy
python-multipart
